--- src/components/comment/List/Comment/CommentDelete/CommentDeleteButton.tsx
@@ -1,12 +1,0 @@
-'use client'
-
-export default function CommentDeleteButton() {
-  return (
-    <button
-      type="button"
-      className="p-2.5 text-xs text-left hover:bg-blue-50 hover:text-blue-600 rounded-[8px]"
-    >
-      댓글 삭제
-    </button>
-  )
-}
